+++ app/src/main/java/com/github/bumblebee202111/doubean/feature/login/navigation/LoginNavigation.kt
@@ -14,7 +14,7 @@ fun NavController.navigateToLogin() = navigate(route = LoginRoute)
 fun NavGraphBuilder.loginScreen(
     onSaveIsLoginSuccessSuccessfulChange: (Boolean) -> Unit,
     onPopBackStack: () -> Unit,
-    onOpenDeepLinkUrl: (url: String) -> Unit,
+    onOpenDeepLinkUrl: (url: String) -> Boolean,
 ) = composable<LoginRoute> {
     LoginScreen(
         onSaveIsLoginSuccessSuccessfulChange = onSaveIsLoginSuccessSuccessfulChange,