+++ app/src/main/java/com/github/bumblebee202111/doubean/model/fangorns/ColorScheme.kt
@@ -1,0 +1,8 @@
+package com.github.bumblebee202111.doubean.model.fangorns
+
+data class ColorScheme(
+    val isDark: Boolean,
+    val primaryColorDark: String,
+    val primaryColorLight: String,
+    val secondColor: String,
+)